+++ src/app/services/drop-files.js
@@ -1,11 +1,7 @@
 'use strict';
 
 const {extOf, fileKind, hashPath, isImageFile} = require('../util/file-tree');
-
-const isElectronBridge = () =>
-	typeof window !== 'undefined'
-	&& window.app
-	&& window.app.platform === 'electron';
+const {isElectronBridge} = require('../util/platform');
 
 const nodeFromPath = filePath => {
 	const name = filePath.split(/[/\\]/).pop() || filePath;